pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stemforge"
version = "0.2.0"
description = "Stem splitting + beat slicing pipeline for Ableton Live IDM production"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
local = ["torch>=2.1.0", "torchaudio>=2.1.0", "demucs>=4.0.1"]

[project.scripts]
stemforge = "stemforge.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["stemforge*"]
